refactor: rename apps & models
