+++ apps/companies/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class CompanysConfig(AppConfig):
+class CompaniesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.companys"
+    name = "apps.companies"
