+++ apps/companies/templates/companies/edit.html
@@ -3,7 +3,7 @@
 {% block content %}
 <h1>編輯公司欄位</h1>
 
-<form method="post" action="{% url 'companys:show' company.id %}">
+<form method="post" action="{% url 'companies:show' company.id %}">
     {% csrf_token %}
     {{ form }}
     <div><button>修改</button></div>
