+++ apps/companies/templates/companies/index.html
@@ -1,0 +1,20 @@
+{% extends "layout/base.html" %}
+
+{% block content %}
+<h1>公司欄位首頁</h1>
+
+<a href="{% url 'companies:new' %}">新增</a>
+
+<ul>
+    {% for company in companies %}
+        <li>
+            <a href="{% url 'companies:show' company.id %}">{{ company.title }}</a>
+            <a href="{% url 'companies:edit' company.id %}">編輯</a>
+            <form method="post" action="{% url 'companies:deleted' company.id %}">
+                {% csrf_token %}
+                <button>刪除</button>
+            </form>
+        </li>
+    {% endfor %}
+</ul>
+{% endblock %}
