+++ apps/companies/templates/companies/show.html
@@ -10,5 +10,5 @@
     <li>{{ company.url }}</li>
 </ul>
 
-<a href="{% url 'companys:index' %}">返回列表</a>
+<a href="{% url 'companies:index' %}">返回列表</a>
 {% endblock %}
