+++ apps/companies/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
+
 from . import views
 
-app_name = "companys"
+app_name = "companies"
 
 urlpatterns = [
     path("", views.index, name="index"),
