--- apps/companys/templates/companys/index.html
@@ -1,20 +1,0 @@
-{% extends "layout/base.html" %}
-
-{% block content %}
-<h1>公司欄位首頁</h1>
-
-<a href="{% url 'companys:new' %}">新增</a>
-
-<ul>
-    {% for company in companys %}
-        <li>
-            <a href="{% url 'companys:show' company.id %}">{{ company.companys_title }}</a>
-            <a href="{% url 'companys:edit' company.id %}">編輯</a>
-            <form method="post" action="{% url 'companys:deleted' company.id %}">
-                {% csrf_token %}
-                <button>刪除</button>
-            </form>
-        </li>
-    {% endfor %}
-</ul>
-{% endblock %}
