+++ config/urls.py
@@ -20,7 +20,7 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path("companys/", include("apps.companys.urls")),
+    path("companies/", include("apps.companies.urls")),
     path("admin/", admin.site.urls),
     path("", include("apps.users.urls")),
     path("jobs/", include("apps.jobs.urls")),
